Edit integer to float

+++ src/Sensor/PressureSensor.cpp
@@ -18,7 +18,7 @@ float PressureSensor::toHz(){
 	return 0;
 }
 
-int PressureSensor::toValue(){
+float PressureSensor::toValue(){
 	if (i2c.transaction(SENSOR_ADDR, &readPressureCmd, 1, pressureData, 3)) {
 		/* Output temperature. */
 		pressure = (pressureData[0] << 8) | pressureData[1];
